split off `build-file` tests from huge `build`
added some target tests to improve test coverage
